refactor(redis): 优化 DBSize 类型和更新连接配置
- 将 DBSize 的返回类型从 int 改为 long，以支持更大的数据库大小
- 在 NewConnectionController 中为 redisConfig 设置 name属性

+++ rdm-redis-imp/src/main/java/xyz/hashdog/rdm/redis/imp/client/JedisClusterClient.java
@@ -116,12 +116,12 @@ public class JedisClusterClient extends AbstractRedisClient implements RedisClie
     public Map<Integer, Integer> dbSize() {
         return execute(jedis->{
             Map<Integer,Integer> map = new LinkedHashMap<>();
-            int dbSize=0;
+            long dbSize=0;
             for (String master : masters) {
                 Connection connection = this.jedis.getClusterNodes().get(master).getResource();
-                dbSize += (int)connection.executeCommand(Protocol.Command.DBSIZE);
+                dbSize += (long)connection.executeCommand(Protocol.Command.DBSIZE);
             }
-            map.put(0,dbSize);
+            map.put(0, (int) dbSize);
             return map;
         });
     }
